# analyses/OpenAPI_ColisApp.yaml
openapi: 3.0.0
info:
  title: ColisApp API
  description: Documentation de l'API RESTful pour l'application de gestion des envois ColisApp.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Serveur de développement local
  - url: https://colisapp.example.com/api/v1
    description: Serveur de production
security:
  - BearerAuth: []
...
paths:
...
  /envois/{id}:
    get:
      summary: Récupérer un envoi par ID
      description: Récupère un enregistrement d'envoi unique par son ID. Nécessite une authentification.
      tags:
        - Envois
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID de l'envoi à récupérer
      responses:
        '200':
          description: Détails de l'envoi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvoiResponseDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Mettre à jour un envoi par ID
      description: Met à jour un enregistrement d'envoi existant par son ID. Nécessite une authentification.
      tags:
        - Envois
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID de l'envoi à mettre à jour
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEnvoiDto'
      responses:
        '200':
          description: Envoi mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvoiDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /agencies/findAgency:
    get:
      summary: Trouver l'ID d'une agence par ses détails
      description: Trouve l'ID d'une agence par son nom, sa ville et son pays.
      tags:
        - Agencies
      parameters:
        - in: query
          name: country
          schema:
            type: integer
          required: true
          description: ID du pays
        - in: query
          name: city
          schema:
            type: string
          required: true
          description: Nom de la ville
        - in: query
          name: agency_name
          schema:
            type: string
          required: true
          description: Nom de l'agence
      responses:
        '200':
          description: ID de l'agence trouvé
          content:
            application/json:
              schema:
                type: object
                properties:
                  agencyId:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /agencies/light:
    get:
      summary: Récupérer une liste allégée d'agences
      description: Récupère une liste allégée d'agences, filtrable par ID de pays, ID de ville et terme de recherche.
      tags:
        - Agencies
      parameters:
        - in: query
          name: countryId
          schema:
            type: integer
          description: ID du pays
        - in: query
          name: cityId
          schema:
            type: integer
          description: ID de la ville
        - in: query
          name: search
          schema:
            type: string
          description: Terme de recherche pour le nom de l'agence
      responses:
        '200':
          description: Liste allégée d'agences
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /agencies/update-agency:
    put:
      summary: Mettre à jour une agence existante
      description: Met à jour un enregistrement d'agence existant dans la base de données.
      tags:
        - Agencies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgencyDto'
      responses:
        '200':
          description: Agence mise à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgencyDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cities/{countryId}:
    get:
      summary: Récupérer les villes par ID de pays
      description: Récupère une liste de toutes les villes pour un ID de pays donné.
      tags:
        - Cities
      parameters:
        - in: path
          name: countryId
          schema:
            type: integer
          required: true
          description: ID du pays
      responses:
        '200':
          description: Liste de villes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /contact:
    post:
      summary: Envoyer un message de contact
      description: Envoie un message via le formulaire de contact.
      tags:
        - Contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageBodyDto'
      responses:
        '200':
          description: Message envoyé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '403':
          description: CORS non autorisé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '500':
          $ref: '#/components/responses/InternalServerError'
    options:
      summary: Gérer les requêtes pré-vol CORS
      description: Gère les requêtes pré-vol (OPTIONS) pour le formulaire de contact.
      tags:
        - Contact
      responses:
        '204':
          description: Pré-vol CORS réussi
  /countries/all:
    get:
      summary: Récupérer tous les pays
      description: Récupère une liste de tous les pays dans le système.
      tags:
        - Countries
      responses:
        '200':
          description: Liste de tous les pays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryDto'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /countries:
    get:
      summary: Récupérer les pays distincts avec agences
      description: Récupère une liste de pays distincts qui ont des agences, avec une option pour exclure un pays de départ spécifique.
      tags:
        - Countries
      parameters:
        - in: query
          name: departureCountry
          schema:
            type: integer
          description: ID du pays à exclure de la liste
      responses:
        '200':
          description: Liste de pays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryResponseDto'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Non autorisé
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unauthorized
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Not found
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Invalid data provided
    MethodNotAllowed:
      description: Méthode non autorisée
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Method not allowed
    InternalServerError:
      description: Erreur interne du serveur
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Internal server error
...
  schemas:
...
    RoleDto:
      type: string
      enum:
        - CLIENT
        - SUPER_ADMIN
        - DESTINATAIRE
        - AGENCY_ADMIN
        - ACCOUNTANT
    SimulationStatusDto:
      type: string
      enum:
        - DRAFT
        - CONFIRMED
        - COMPLETED
        - CANCELLED
    EnvoiStatusDto:
      type: string
      enum:
        - PENDING
        - SENT
        - DELIVERED
        - CANCELLED
        - RETURNED
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        name:
          type: string
        birthDate:
          type: string
          format: date
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        password:
          type: string
          format: password
        image:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/RoleDto'
        agencyId:
          type: integer
        isVerified:
          type: boolean
        emailVerified:
          type: string
          format: date-time
        verificationToken:
          type: string
        verificationTokenExpires:
          type: string
          format: date-time
        userAddresses:
          $ref: '#/components/schemas/UserAddressDto'
          nullable: true
      required:
        - email
...
    AgencyDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        vatNumber:
          type: string
        addressId:
          type: integer
        address:
          $ref: '#/components/schemas/AddressDto'
        capacity:
          type: integer
          nullable: true
        availableSlots:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - name
        - addressId
        - createdAt
        - updatedAt
...
    UpdateAgencyDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        vatNumber:
          type: string
        addressId:
          type: integer
        address:
          $ref: '#/components/schemas/AddressDto'
        capacity:
          type: integer
          nullable: true
        availableSlots:
          type: integer
          nullable: true
      required:
        - id

    # Envoi DTOs
    EnvoiDto:
      type: object
      properties:
        id:
          type: integer
        destinationId:
          type: integer
        trackingNumber:
          type: string
        qrCodeUrl:
          type: string
        userId:
          type: integer
        client:
          $ref: '#/components/schemas/UserDto'
        destinataireId:
          type: integer
        destinataire:
          $ref: '#/components/schemas/UserDto'
        transportId:
          type: integer
        departureAgencyId:
          type: integer
        departureAgency:
          $ref: '#/components/schemas/AgencyDto'
          nullable: true
        arrivalAgencyId:
          type: integer
        arrivalAgency:
          $ref: '#/components/schemas/AgencyDto'
          nullable: true
        simulationStatus:
          $ref: '#/components/schemas/SimulationStatusDto'
        envoiStatus:
          $ref: '#/components/schemas/EnvoiStatusDto'
        paid:
          type: boolean
        totalWeight:
          type: number
          format: float
        totalVolume:
          type: number
          format: float
        totalPrice:
          type: number
          format: float
        departureDate:
          type: string
          format: date-time
        arrivalDate:
          type: string
          format: date-time
        verificationToken:
          type: string
        comment:
          type: string
        parcels:
          type: array
          items:
            $ref: '#/components/schemas/ParcelDto'
        createdAt:
          type: string
          format: date-time
      required:
        - departureAgencyId
        - arrivalAgencyId
        - simulationStatus
        - envoiStatus
        - paid
        - totalWeight
        - totalVolume
        - totalPrice
        - departureDate
        - arrivalDate
        - verificationToken
...
    UpdateEnvoiDto:
      type: object
      properties:
        id:
          type: integer
        destinationId:
          type: integer
        trackingNumber:
          type: string
        qrCodeUrl:
          type: string
        userId:
          type: integer
        destinataireId:
          type: integer
        transportId:
          type: integer
        departureAgencyId:
          type: integer
        arrivalAgencyId:
          type: integer
        simulationStatus:
          $ref: '#/components/schemas/SimulationStatusDto'
        envoiStatus:
          $ref: '#/components/schemas/EnvoiStatusDto'
        paid:
          type: boolean
        totalWeight:
          type: number
          format: float
        totalVolume:
          type: number
          format: float
        totalPrice:
          type: number
          format: float
        departureDate:
          type: string
          format: date-time
        arrivalDate:
          type: string
          format: date-time
        verificationToken:
          type: string
        comment:
          type: string
        parcels:
          type: array
          items:
            $ref: '#/components/schemas/ParcelDto'
        createdAt:
          type: string
          format: date-time
      required:
        - id
    EnvoiResponseDto:
      type: object
      properties:
        id:
          type: integer
        destinataire:
          type: object
          properties:
            name:
              type: string
            email:
              type: string
              format: email
            phoneNumber:
              type: string
          required:
            - name
            - email
            - phoneNumber
        client:
          $ref: '#/components/schemas/UserDto'
        trackingNumber:
          type: string
        qrCodeUrl:
          type: string
        arrivalAgencyId:
          type: integer
        departureAgencyId:
          type: integer
        envoiStatus:
          $ref: '#/components/schemas/EnvoiStatusDto'
        simulationStatus:
          $ref: '#/components/schemas/SimulationStatusDto'
        userId:
          type: integer
        paid:
          type: boolean
        transportId:
          type: integer
        parcels:
          type: array
          items:
            $ref: '#/components/schemas/ParcelDto'
        totalWeight:
          type: number
          format: float
        totalVolume:
          type: number
          format: float
        totalPrice:
          type: number
          format: float
        departureDate:
          type: string
          format: date-time
        arrivalDate:
          type: string
          format: date-time
      required:
        - id
        - destinataire
        - client
        - trackingNumber
        - qrCodeUrl
        - arrivalAgencyId
        - departureAgencyId
        - envoiStatus
        - simulationStatus
        - userId
        - paid
        - transportId
        - parcels
        - totalWeight
        - totalVolume
        - totalPrice
        - departureDate
        - arrivalDate

    # Parcel DTOs
    ParcelDto:
      type: object
      properties:
        id:
          type: integer
        height:
          type: number
          format: float
        weight:
          type: number
          format: float
        width:
          type: number
          format: float
        length:
          type: number
          format: float
        envoiId:
          type: integer
      required:
        - height
        - weight
        - width
        - length
        - envoiId
...
    AddressDto:
      type: object
      properties:
        id:
          type: integer
        street:
          type: string
        complement:
          type: string
          nullable: true
        streetNumber:
          type: string
          nullable: true
        boxNumber:
          type: string
          nullable: true
        cityId:
          type: integer
        city:
          $ref: '#/components/schemas/CityDto'
      required:
        - street
        - city
    UserAddressDto:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        addressId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        addressType:
          $ref: '#/components/schemas/AddressType'
        isDeleted:
          type: boolean
        address:
          $ref: '#/components/schemas/AddressDto'
        user:
          $ref: '#/components/schemas/UserDto'
        city:
          $ref: '#/components/schemas/CityDto'
      required:
        - userId
        - addressId
        - createdAt
        - updatedAt
        - addressType
        - isDeleted
        - address
        - user
        - city
...
    CityDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          $ref: '#/components/schemas/CountryDto'
      required:
        - id
        - name
        - country

    # Country DTOs
    CountryDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        iso2:
          type: string
          nullable: true
        iso3:
          type: string
          nullable: true
        phonecode:
          type: string
          nullable: true
        capital:
          type: string
          nullable: true
        currency:
          type: string
          nullable: true
        latitude:
          type: string
          nullable: true
        longitude:
          type: string
          nullable: true
      required:
        - id
        - name